Dialogue System
I got the interaction system fully operational. I got the dialogue system in and you can talk to the cube in the scene. I am going to be working on using xml to make the dialogue system so you can have options to choose from and create the branching dialogue system