Scriptable Objects
I got the scriptable objects all set up for items and events. An NPC entering an item trigger causes the event to be called. I have listeners in the scene that respond to the events. Pretty cool! I am working on implementing it with the NPCs that Will set up yesterday.